third_party/blink/web_tests/accessibility/aria-row-name.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<body id="body">
<div role="treegrid">
  <div role="row" id="treegrid1-row1" tabindex="-1">
    <div role="rowheader">row head</div>
    <div role="gridcell">data</div>
  </div>
</div>
<div role="treegrid" aria-activedescendant="treegrid2-row1">
  <div role="row" id="treegrid2-row1">
    <div role="rowheader">row head</div>
    <div role="gridcell">data</div>
  </div>
</div>
<div role="grid">
  <div role="row" id="treegrid3-row1">
    <div role="rowheader">row head</div>
    <div role="gridcell">data</div>
  </div>
</div>

<input type="text" role="combobox" aria-activedescendant="treegrid4-row1">
<div role="grid">
    <div id="treegrid4-row1" role="row">
        <div role="gridcell">data</div>
    </div>
</div>
</body>

<input type="text" role="combobox" aria-controls="treegrid5" aria-activedescendant="treegrid5-row1">
<div id="treegrid5" role="grid">
    <div id="treegrid5-row1" role="row">
        <div role="gridcell">data</div>
    </div>
</div>
</body>

<script>
function axElementById(id) {
    return accessibilityController.accessibleElementById(id);
}

test(function(t) {
    var axRow = axElementById("treegrid1-row1");
    assert_equals(axRow.name, 'row head data');
}, "The row's name must concatenate the children if the row is focusable");

test(function(t) {
    var axRow = axElementById("treegrid2-row1");
    assert_equals(axRow.name, 'row head data');
}, "The row's name must concatenate the children if the parent has aria-activedescendant");

test(function(t) {
    var axRow = axElementById("treegrid3-row1");
    assert_equals(axRow.name, '');
}, "The row's name must empty when the parent is not focusable and there is no other labelling markup (for performance reasons)");

test(function(t) {
    var axRow = axElementById("treegrid4-row1");
    assert_equals(axRow.name, 'data');
}, "The row's name must concatenate the children if a textfield controlling the grid has aria-activedescendant");

test(function(t) {
    var axRow = axElementById("treegrid5-row1");
    assert_equals(axRow.name, 'data');
}, "The row's name must concatenate the children if the grid has a previous sibling input with aria-activedescendant");

</script>
